Removed modules from airframe conf file. Changed datatypes in inter_thread_data.h. Done some work on the opticflow calculations.

+++ sw/airborne/modules/computer_vision/opticflow/inter_thread_data.h
@@ -35,13 +35,13 @@ struct opticflow_result_t {
   uint16_t corner_cnt;    //< The amount of coners found by FAST9
   uint16_t tracked_cnt;   //< The amount of tracked corners
 
-  int16_t flow_x;         //< Flow in x direction from the camera (in subpixels)
+  float flow_x;         //< Flow in x direction from the camera (in subpixels)
   int16_t flow_y;         //< Flow in y direction from the camera (in subpixels)
-  int16_t flow_der_x;     //< The derotated flow calculation in the x direction (in subpixels)
+  float flow_der_x;     //< The derotated flow calculation in the x direction (in subpixels)
   
-  int16_t tot_of_left;	  // Total optic flow left
-  int16_t tot_of_right;	  // Total optic flow right
-  int16_t tot_of;	  // Total optic flow
+  float tot_of_left;	  // Total optic flow left
+  float tot_of_right;	  // Total optic flow right
+  float tot_of;	  // Total optic flow
 };
 
 /* The state of the drone when it took an image */
